pdf-tools/img-to-pdf/requirements.txt
pillow==12.0.0
PyPDF2==3.0.1
